// src/types/api/service/models/MailTemplatesFooterDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MailTemplatesFooterDto
 */
export interface MailTemplatesFooterDto {
    /**
     * 
     * @type {number}
     * @memberof MailTemplatesFooterDto
     */
    mailTemplatesFooterId?: number;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    mailWhy?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    senderEmail?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    senderName?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    format?: string;
    /**
     * 
     * @type {string}
     * @memberof MailTemplatesFooterDto
     */
    language?: string;
}

/**
 * Check if a given object implements the MailTemplatesFooterDto interface.
 */
export function instanceOfMailTemplatesFooterDto(value: object): value is MailTemplatesFooterDto {
    return true;
}

export function MailTemplatesFooterDtoFromJSON(json: any): MailTemplatesFooterDto {
    return MailTemplatesFooterDtoFromJSONTyped(json, false);
}

export function MailTemplatesFooterDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MailTemplatesFooterDto {
    if (json == null) {
        return json;
    }
    return {
        
        'mailTemplatesFooterId': json['mailTemplatesFooterId'] == null ? undefined : json['mailTemplatesFooterId'],
        'mailWhy': json['mailWhy'] == null ? undefined : json['mailWhy'],
        'title': json['title'] == null ? undefined : json['title'],
        'body': json['body'] == null ? undefined : json['body'],
        'senderEmail': json['senderEmail'] == null ? undefined : json['senderEmail'],
        'senderName': json['senderName'] == null ? undefined : json['senderName'],
        'format': json['format'] == null ? undefined : json['format'],
        'language': json['language'] == null ? undefined : json['language'],
    };
}

export function MailTemplatesFooterDtoToJSON(value?: MailTemplatesFooterDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'mailTemplatesFooterId': value['mailTemplatesFooterId'],
        'mailWhy': value['mailWhy'],
        'title': value['title'],
        'body': value['body'],
        'senderEmail': value['senderEmail'],
        'senderName': value['senderName'],
        'format': value['format'],
        'language': value['language'],
    };
}